% Generated by roxygen2 (4.0.2): do not edit by hand
\name{effect.Rchoice}
\alias{effect.Rchoice}
\title{Get the conditional individual coefficients}
\usage{
effect.Rchoice(x, par = NULL, effect = c("cv", "ce"), wrt = NULL, ...)
}
\arguments{
\item{x}{a object of class \code{Rchoice},}

\item{par}{a string giving the name of the variable with random parameter,}

\item{effect}{a string indicating what should be plotted: the conditional expectation of the individual coefficients "\code{ce}", or the conditional expectation of the individual compensating variations "\code{cv}",}

\item{wrt}{a string indicating repect to which variable should be computed the compensating variation,}

\item{...}{further arguments. Ignored.}
}
\description{
This a helper function to obtain the conditional estimate of the individual random parameters or the compensating variations.
}

