% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetprob.R
\name{getSummary.effect.hetprob}
\alias{getSummary.effect.hetprob}
\title{Get Model Summaries for use with "mtable" for objects of class effect.hetprob}
\usage{
\method{getSummary}{effect.hetprob}(obj, alpha = 0.05, ...)
}
\arguments{
\item{obj}{an \code{effect.hetprob} object,}

\item{alpha}{level of the confidence intervals,}

\item{...}{further arguments,}
}
\description{
A generic function to collect coefficients and summary statistics from a \code{effect.hetprob} object. It is used in \code{mtable}
}
\details{
For more details see package \pkg{memisc}.
}
