% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivpml.R
\name{getSummary.effect.ivpml}
\alias{getSummary.effect.ivpml}
\title{Get Model Summaries for use with "mtable" for objects of class effect.ivpml}
\usage{
\method{getSummary}{effect.ivpml}(obj, alpha = 0.05, ...)
}
\arguments{
\item{obj}{an \code{effect.ivpml} object,}

\item{alpha}{level of the confidence intervals,}

\item{...}{further arguments,}
}
\description{
A generic function to collect coefficients and summary statistics from a \code{effect.ivpml} object. It is used in \code{mtable}
}
\details{
For more details see package \pkg{memisc}.
}
