% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSummary.Rchoice}
\alias{getSummary.Rchoice}
\title{Get Model Summaries for Use with "mtable"}
\usage{
getSummary.Rchoice(obj, alpha = 0.05, ...)
}
\arguments{
\item{obj}{a \code{Rchoice} object,}

\item{alpha}{level of the confidence intervals,}

\item{...}{further arguments,}
}
\description{
A generic function to collect coefficients and summary statistics from a \code{Rchoice} object. It is used in \code{mtable}
}

