% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rchoice.modelform.R
\name{rFormula}
\alias{is.rFormula}
\alias{model.frame.rFormula}
\alias{model.matrix.rFormula}
\alias{rFormula}
\title{Model formula for Rchoice models}
\usage{
rFormula(object)

is.rFormula(object)

\method{model.frame}{rFormula}(formula, data, ..., lhs = NULL, rhs = NULL)

\method{model.matrix}{rFormula}(object, data, rhs = NULL, ...)
}
\arguments{
\item{object}{a formula form the \code{rFormula} function, for the \code{model.matrix} method, a \code{rFormula} object,}

\item{formula}{a \code{rFormula} object,}

\item{data}{a \code{data.frame},}

\item{...}{further arguments.}

\item{lhs}{see \code{\link[Formula]{Formula}},}

\item{rhs}{see \code{\link[Formula]{Formula}},}
}
\description{
Two kind of variables are used in models with individual heterogenetiy: the typical
variables that enter in the latent process and those variables that enter in the random
parameter (Hierarchical Model). \code{rFormula} deal with this type of models using
suitable methods to extract the elements of the model.
}

