% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rchoice.methods.R
\name{estfun.Rchoice}
\alias{estfun.Rchoice}
\title{Gradient for observations}
\usage{
\method{estfun}{Rchoice}(x, ...)
}
\arguments{
\item{x}{a fitted model of class \code{Rchoice},}

\item{...}{Other arguments when \code{estfun} is applied to another
class object}
}
\value{
the gradient matrix of dimension n times k
}
\description{
It extracts the gradient for each observations evaluated at the estimated parameters for a model of class \code{Rchoice}
}
\details{
For more information see \code{\link[sandwich]{estfun}} from package \pkg{sandwich}.
}
\references{
Zeileis A (2006), Object-oriented Computation of Sandwich 
Estimators. Journal of Statistical Software, 16(9), 1--16.
}
