% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rchoice.methods.R
\name{getSummary.Rchoice}
\alias{getSummary.Rchoice}
\title{Get Model Summaries for Use with "mtable"}
\usage{
getSummary.Rchoice(obj, alpha = 0.05, ...)
}
\arguments{
\item{obj}{a \code{Rchoice} object,}

\item{alpha}{level of the confidence intervals,}

\item{...}{further arguments,}
}
\description{
A generic function to collect coefficients and summary statistics from a \code{Rchoice} object. It is used in \code{mtable}
}
\details{
For more details see package \pkg{memisc}.
}
\examples{
## Probit Model
data("Workmroz")
probit <- Rchoice(lfp ~ k5 + k618 + age + wc + hc + lwg + inc,
                 data = Workmroz, family = binomial('probit'))
## Logit Model
logit <- Rchoice(lfp ~ k5 + k618 + age + wc + hc + lwg + inc,
                 data = Workmroz, family = binomial('logit'))

## Table with Models
library(memisc)
mtable("Probit Model"= probit, "Logit Model" = logit,
       summary.stats = c("N", "Log-likelihood", "BIC", "AIC"))
}

