% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rchoice.methods.R
\name{getSummary.Rchoice}
\alias{getSummary.Rchoice}
\title{Get Model Summaries for use with "mtable" for object of class Rchoice}
\usage{
\method{getSummary}{Rchoice}(obj, alpha = 0.05, ...)
}
\arguments{
\item{obj}{a \code{Rchoice} object,}

\item{alpha}{level of the confidence intervals,}

\item{...}{further arguments,}
}
\description{
A generic function to collect coefficients and summary statistics from a \code{Rchoice} object. It is used in \code{mtable}
}
\details{
For more details see package \pkg{memisc}.
}
